Write Hono client config files to dedicated folder
The property files used to configure the MQTT and Kafka
clients of the FMS Forwarder and FMS Consumer are now
created in a dedicated folder.

The FMS Forwarder now also (bind) mounts a folder from that
location which may contain client keys and trusted
certs for connecting to the MQTT adapter.