Do not use default VIN
The default VIN had been used when the Databroker did not contain a
VIN (yet). This was often the case when the CSV provider did not
report the VIN to the Databroker soon after startup but only later.

This had the unwanted effect of two VINs showing up in the influxDB,
the default VIN and (later) the real VIN as reported by the CSV
provider.

The FMS Forwarder has been changed to ignore data sets from the
Databroker unless it contains a VIN.